+++ apps/backoffice-v2/src/pages/Entity/components/EditableDetails/EditableDetails.tsx
@@ -24,6 +24,7 @@ import { keyFactory } from '../../../../common/utils/key-factory/key-factory';
 import { isObject } from '@ballerine/common';
 import { isValidUrl } from '../../../../common/utils/is-valid-url';
 import { JsonDialog } from '../../../../common/components/molecules/JsonDialog/JsonDialog';
+import { FileJson2 } from 'lucide-react';
 
 const useInitialCategorySetValue = ({ form, data }) => {
   useEffect(() => {
@@ -42,8 +43,14 @@ export const EditableDetails: FunctionComponent<IEditableDetails> = ({
   workflowId,
 }) => {
   const [formData, setFormData] = useState(data);
+  const useInitialCategorySetValue = () => {
+    useEffect(() => {
+      const categoryValue = form.getValues('category');
+      form.setValue('category', categoryValue);
+    }, [form, data, setFormData]);
+  };
   const POSITIVE_VALUE_INDICATOR = ['approved'];
-  const NEGATIVE_VALUE_INDICATOR = ['revision', 'rejected'];
+  const NEGATIVE_VALUE_INDICATOR = ['revision', 'rejected', 'declined'];
   const isDecisionPositive = (isDecisionComponent: boolean, value: string) => {
     return isDecisionComponent && value && POSITIVE_VALUE_INDICATOR.includes(value.toLowerCase());
   };
@@ -109,10 +116,10 @@ export const EditableDetails: FunctionComponent<IEditableDetails> = ({
     data,
   });
 
-  // Ensures that the form is reset when the data changes from other instances of `useUpdateWorkflowByIdMutation` i.e. in `useCallToActionLogic`.
+  // Ensures that the form is reset when the data changes from other instances of `useUpdateWorkflowByIdMutation` i.e. in `useCaseCallToActionLogic`.
   useEffect(() => {
     form.reset(defaultValues);
-  }, [form.reset, defaultValues]);
+  }, [form.reset, data]);
 
   return (
     <Form {...form}>
@@ -151,6 +158,11 @@ export const EditableDetails: FunctionComponent<IEditableDetails> = ({
                             key={keyFactory(valueId, title, `form-field`)}
                           >
                             <JsonDialog
+                              buttonProps={{
+                                variant: 'link',
+                                className: 'p-0 text-blue-500',
+                              }}
+                              rightIcon={<FileJson2 size={`16`} />}
                               dialogButtonText={`View Information`}
                               json={JSON.stringify(value)}
                             />