+++ packages/workflow-core/src/lib/plugins/external-plugin/kyc-session-plugin.ts
@@ -23,20 +23,6 @@ const kycSessionRequestSchema = {
   required: ['firstName', 'lastName', 'callbackUrl', 'vendor'],
 };
 
-const kycSessionResponseSchema = {
-  $schema: 'http://json-schema.org/draft-07/schema#',
-  type: 'object',
-  properties: {
-    id: {
-      type: 'string',
-    },
-    url: {
-      type: 'string',
-    },
-  },
-  required: ['id', 'url'],
-};
-
 export class KycSessionPlugin extends ApiPlugin {
   public static pluginType = 'http';
   public static pluginKind = 'kyc-session';
@@ -46,11 +32,14 @@ export class KycSessionPlugin extends ApiPlugin {
     transformedRequest: AnyRecord,
     validationContext: TValidationContext,
   ) {
-    return super.validateContent(
-      new JsonSchemaValidator(validationContext === 'Request' ? kycSessionRequestSchema : kycSessionResponseSchema),
-      transformedRequest,
-      validationContext,
-    );
+    if (validationContext === 'Request') {
+      return super.validateContent(
+        new JsonSchemaValidator(kycSessionRequestSchema),
+        transformedRequest,
+        validationContext,
+      );
+    }
+    return super.validateContent(schemaValidator, transformedRequest, validationContext);
   }
 
   async makeApiRequest(