+++ services/workflows-service/scripts/workflows/kyc-email-process-example.ts
@@ -27,15 +27,32 @@ export const kycEmailSessionDefinition = {
         },
       },
       kyc_manual_review: {
-        type: 'final' as const,
+        on: {
+          approve: {
+            target: 'approved',
+          },
+          reject: {
+            target: 'rejected',
+          },
+          revision: {
+            target: 'revision',
+          },
+        },
+      },
+      revision: {
+        always: [
+          {
+            target: 'get_kyc_session',
+          },
+        ],
       },
       kyc_auto_reject: {
         type: 'final' as const,
       },
-      reject: {
+      rejected: {
         type: 'final' as const,
       },
-      approve: {
+      approved: {
         type: 'final' as const,
       },
     },
@@ -96,7 +113,8 @@ export const kycEmailSessionDefinition = {
               receivers: [entity.data.email],
               subject: '{customerCompanyName} activation, Action needed.',
               preheader: 'Verify your identity for Happy Home Goods activation with {customerCompanyName}.',
-              templateId: 'd-61c568cfa5b145b5916ff89790fe2065'
+              templateId: 'd-61c568cfa5b145b5916ff89790fe2065',
+              revisionReason: documents[].decision[].revisionReason | [0]
               }`, // jmespath
             },
           ],